requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.11.12
charset-normalizer==3.4.4
dnspython==2.8.0
feedparser==6.0.12
filelock==3.20.0
frozenlist==1.8.0
geoip2==5.1.0
idna==3.11
markdown-it-py==4.0.0
maxminddb==2.8.2
mdurl==0.1.2
motor==3.7.1
multidict==6.7.0
numpy==2.3.4
orjson==3.11.4
pandas==2.3.3
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.0.1
requests==2.32.5
requests-file==3.0.1
rich==14.2.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8
tldextract==5.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
yarl==1.22.0
stix2>=3.0.0
